test: split caption-bounded PDF panel coverage
